added catch for error and error message, fix bug with rent counting the cards rented out
added catch in loading/reloading accounts to fix the issue of tracker stopping to get the other accounts details when there was an issue with an account. The tracker will no1 show the issue on the top page before the dashboard
Fix the issue in net income where the cards rent out by the user is being counter in the rent per day.